Fixed translucents

+++ shaders/bufferSettings.glsl
@@ -8,7 +8,7 @@
 
 /*
 MAIN BUFFER
-const int colortex0Format = RGB32F;
+const int colortex0Format = RGB16F;
 const bool colortex0Clear = false;
 
 DATA BUFFER (SOLID)
